feat: `template` に書かれていたロジックを `Header` に移動
